+++ src/http/device_http.h
@@ -1,0 +1,17 @@
+/*
+ * device_http.h
+ *
+ *  Created on: 2016/11/28
+ *      Author: lc
+ */
+#ifndef SRC_DEVICE_HTTP_H_
+#define SRC_DEVICE_HTTP_H_
+
+#include <evhttp.h>
+
+void http_deviceHandler(struct evhttp_request *req);
+
+
+#endif/* SRC_DEVICE_HTTP_H_ */
+
+
