+++ src/http/gps_http.h
@@ -1,0 +1,17 @@
+/*
+ * gps_http.h
+ *
+ *  Created on: 2016/11/28
+ *      Author: lc
+ */
+#ifndef SRC_GPS_HTTP_H_
+#define SRC_GPS_HTTP_H_
+
+#include <evhttp.h>
+
+void http_replyGPS(struct evhttp_request *req);
+
+
+
+#endif/* SRC_GPS_HTTP_H_ */
+
