+++ src/http/itinerary_http.h
@@ -1,0 +1,17 @@
+/*
+ * itinerary_http.h
+ *
+ *  Created on: 2016/11/28
+ *      Author: lc
+ */
+#ifndef SRC_ITINERARY_HTTP_H_
+#define SRC_ITINERARY_HTTP_H_
+
+#include <evhttp.h>
+
+void http_replyItinerary(struct evhttp_request *req);
+
+
+
+#endif/* SRC_ITINERARY_HTTP_H_ */
+
