+++ src/http/main.c
@@ -30,6 +30,9 @@ static void signal_handler(int sig)
 
 int main(int argc, char *argv[])
 {
+
+	struct evhttp_bound_socket *handle;
+
     int http_port = PORT_HTTP;
     char *httpd_listen = "0.0.0.0"; // locaol adress
     int httpd_timeout = 30;         // in seconds
@@ -38,8 +41,8 @@ int main(int argc, char *argv[])
 
     if (argc >= 2)
     {
-    	char* strPort = argv[1];
-    	int num = atoi(strPort);
+    	char* port = argv[1];
+    	int num = atoi(port);
     	if (num)
     	{
     		http_port = num;
@@ -60,24 +63,50 @@ int main(int argc, char *argv[])
         return -1;
     }
 
-    signal(SIGTERM, signal_handler);
-    signal(SIGINT, signal_handler);
+	struct event_base *base = event_base_new();
+	if (!base)
+    {
+		LOG_ERROR("event_base_new failed...");
+		return 1;
+	}
 
-    event_init();
+    struct event *evTerm = evsignal_new(base, SIGTERM, signal_handler, base);
+    if (!evTerm || evsignal_add(evTerm, NULL) < 0)
+    {
+        LOG_ERROR("can't create SIGTERM event");
+    }
 
-    struct evhttp *httpd;
-    httpd = evhttp_start(httpd_listen, http_port);
+    struct event *evInt = evsignal_new(base, SIGINT, signal_handler, base);
+    if (!evInt || evsignal_add(evInt, NULL) < 0)
+    {
+        LOG_ERROR("can't create SIGINT event");
+    }
+
+	struct evhttp *httpd = evhttp_new(base);
+	if (!httpd) {
+		LOG_ERROR("couldn't create evhttp. Exiting.");
+		return 1;
+	}
+
+    if (evhttp_bind_socket(httpd, "0.0.0.0", http_port) != 0)
+    {
+        LOG_ERROR("bind socket failed at port:%d", http_port);
+        return 1;
+    }
+
     evhttp_set_timeout(httpd, httpd_timeout);
-
     evhttp_set_gencb(httpd, httpd_handler, NULL);
 
-    LOG_INFO("start http_server...");
-    event_dispatch();
 
+    LOG_INFO("start http_server: %d", http_port);
+
+    event_base_dispatch(base);
+
+    LOG_INFO("stop http_server: %d", http_port);
+
+    event_base_free(base);
     db_destruct();
-
     zlog_fini();
-
     evhttp_free(httpd);
     return 0;
 }
