+++ src/http/msg_http.h
@@ -9,7 +9,6 @@
 
 #include <evhttp.h>
 
-
 void http_rspMsg(struct evhttp_request *req, char *data);
 void http_errorMsg(struct evhttp_request *req);
 void http_okMsg(struct evhttp_request *req);
